A ver, vuelve posicion a int, y el delta sigue siendo float, pero cuando se va para una posicion se castea el resultado a int...

+++ TP/src/Manual.cpp
@@ -31,7 +31,7 @@ void Manual::moverALaIzquierda(){
 
 void Manual::trasladar(int factor){
 	float DeltaX = velocidad * factor * delta;
-	Posicion* posDesplazamiento = new Posicion (DeltaX,0);
+	Posicion* posDesplazamiento = new Posicion ((int) DeltaX,0);
 	superficieOcupada->mover(posDesplazamiento);
 	delete(posDesplazamiento);
 	huboCambios(); //el método se hereda de Observable
